+++ src/HomePage.tsx
@@ -5,6 +5,6 @@
 // ];
 export const HomePage = () => (
   <div className="container">
-    <h1 className="title">Home page</h1>;
+    <h1 className="title">Home page</h1>
   </div>
 );
