+++ src/Root.tsx
@@ -3,6 +3,12 @@ import { App } from './App';
 import { TabsPage } from './TabsPage';
 import { HomePage } from './HomePage';
 
+const tabs = [
+  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+];
+
 export const Root = () => (
   <HashRouter>
     <Routes>
@@ -11,8 +17,8 @@ export const Root = () => (
         <Route path="home" element={<Navigate to="/" replace />} />
 
         <Route path="tabs">
-          <Route index element={<TabsPage />} />
-          <Route path=":tabId" element={<TabsPage />} />
+          <Route index element={<TabsPage tabs={tabs} />} />
+          <Route path=":tabId" element={<TabsPage tabs={tabs} />} />
         </Route>
         <Route path="*" element={<h1 className="title">Page not found</h1>} />
       </Route>
