+++ src/TabsPage.tsx
@@ -1,33 +1,34 @@
 import { Link, useParams } from 'react-router-dom';
+import { Tabs } from './lib/Tabs';
+import { Tab } from './types/Tab';
+import React from 'react';
 
-export const TabsPage = () => {
-  const { tabId } = useParams();
+type Props = {
+  tabs: Tab[];
+};
 
-  const tabs = [
-    { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-    { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-    { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-  ];
+export const TabsPage: React.FC<Props> = ({ tabs }) => {
+  const { tabId } = useParams<{ tabId: string }>();
 
   const activeTab = tabs.find(tab => tab.id === tabId);
 
   return (
     <div className="container">
       <h1 className="title">Tabs page</h1>
-
-      <div className="tabs is-boxed">
-        <ul>
-          {tabs.map(tab => (
-            <li
-              key={tab.id}
+      <Tabs
+        tabs={tabs.map(tab => ({
+          id: tab.id,
+          title: (
+            <Link
+              to={`/tabs/${tab.id}`}
               className={tab.id === tabId ? 'is-active' : ''}
-              data-cy="Tab"
             >
-              <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
-            </li>
-          ))}
-        </ul>
-      </div>
+              {tab.title}
+            </Link>
+          ),
+        }))}
+        activeTabId={tabId || ''}
+      />
 
       <div className="block" data-cy="TabContent">
         {activeTab ? activeTab.content : 'Please select a tab'}
